Update Tests & add custom render
