+++ src/components/__tests__/units/connect.test.tsx
@@ -1,18 +1,13 @@
 import { Connect } from '@/components/screens/Connect'
 import useAnalytics from '@/hooks/common/useAnalytics'
-import { render, renderHook, screen } from '@testing-library/react'
-import { MockTheme } from '@/tests-tools/theme.mock'
+import { render, screen, renderHook } from '@/tests-tools/test.utils'
 import userEvent from '@testing-library/user-event'
 const onConnectMock = jest.fn()
 
 describe('Connect Screen', () => {
 	it('Page should appears', () => {
 		renderHook(() => useAnalytics())
-		render(
-			<MockTheme>
-				<Connect onConnect={onConnectMock} />
-			</MockTheme>,
-		)
+		render(<Connect onConnect={onConnectMock} />)
 		const scan = screen.getByTestId('scan-button')
 		const input = screen.getByTestId('input-uri')
 		const connectButton = screen.getByTestId('connect-button')
@@ -23,11 +18,7 @@ describe('Connect Screen', () => {
 	it('Page should appears with Connect button and on click triggers action', async () => {
 		renderHook(() => useAnalytics())
 		const url = 'https://jestjs.io/docs/jest-object'
-		render(
-			<MockTheme>
-				<Connect onConnect={onConnectMock} mode="text" />
-			</MockTheme>,
-		)
+		render(<Connect onConnect={onConnectMock} mode="text" />)
 		const connect = screen.getByTestId('connect-button')
 		const input = screen.getByTestId('input-uri')
 		await userEvent.type(input, url)
