--- src/tests-tools/theme.mock.tsx
@@ -1,13 +1,0 @@
-import { StyleProvider } from '@ledgerhq/react-ui'
-
-type PropsTheme = {
-	children: React.ReactElement
-	theme?: 'dark' | 'light'
-}
-export function MockTheme({ children, theme = 'dark' }: PropsTheme) {
-	return (
-		<StyleProvider selectedPalette={theme} fontsPath="/fonts">
-			{children}
-		</StyleProvider>
-	)
-}
